requirements_win.txt
ffmpeg-python
omegaconf
onnx
numpy
opencv-python
gradio
scikit-image
insightface
huggingface_hub[cli]
mediapipe
torchgeometry
soundfile